@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <Nullable>disable</Nullable>
     </PropertyGroup>
 
@@ -8,6 +8,6 @@
       <PackageReference Include="Jellyfin.Common" Version="10.*-*" />
       <PackageReference Include="Jellyfin.Controller" Version="10.*-*" />
       <PackageReference Include="Jellyfin.Extensions" Version="10.*-*" />
-      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.1" />
+      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.8" />
     </ItemGroup>
 </Project>
